Move response type into model.rs
