+++ src/fetcher.rs
@@ -1,6 +1,6 @@
+use crate::model::{KonachanApiResponse, MjxApiPossibleReponse};
 use anyhow::Context;
 use rand::Rng;
-use serde::Deserialize;
 use std::time::Duration;
 
 use crate::data::{DataFetcher, Sendable};
@@ -49,34 +49,6 @@ impl HttpClient {
                     std::any::type_name::<T>()
                 )
             })
-    }
-}
-
-/// Represent the konachan API response json
-#[derive(Deserialize, Debug)]
-pub struct KonachanApiResponse {
-    pub jpeg_url: String,
-    pub file_url: String,
-    pub file_size: u32,
-    pub author: String,
-}
-
-/// The response from MJX API is different. This type can match those different response.
-/// And its associate function can help extract the image link from response.
-#[derive(Deserialize)]
-#[serde(untagged)]
-pub enum MjxApiPossibleReponse {
-    Uomg { code: u8, imgurl: String },
-    Vvhan { title: String, pic: String },
-}
-
-impl MjxApiPossibleReponse {
-    /// Extract the image url from response
-    pub fn unwrap_url(self) -> String {
-        match self {
-            Self::Uomg { imgurl, .. } => imgurl,
-            Self::Vvhan { pic, .. } => pic,
-        }
     }
 }
