+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod data;
 pub mod fetcher;
 pub mod helper;
+pub mod model;
