+++ src/model/mod.rs
@@ -1,0 +1,29 @@
+use serde::Deserialize;
+
+/// Represent the konachan API response json
+#[derive(Deserialize, Debug)]
+pub struct KonachanApiResponse {
+    pub jpeg_url: String,
+    pub file_url: String,
+    pub file_size: u32,
+    pub author: String,
+}
+
+/// The response from MJX API is different. This type can match those different response.
+/// And its associate function can help extract the image link from response.
+#[derive(Deserialize)]
+#[serde(untagged)]
+pub enum MjxApiPossibleReponse {
+    Uomg { code: u8, imgurl: String },
+    Vvhan { title: String, pic: String },
+}
+
+impl MjxApiPossibleReponse {
+    /// Extract the image url from response
+    pub fn unwrap_url(self) -> String {
+        match self {
+            Self::Uomg { imgurl, .. } => imgurl,
+            Self::Vvhan { pic, .. } => pic,
+        }
+    }
+}
